Arreglo carpetas e indentaciones
