+++ src/App.jsx
@@ -3,28 +3,23 @@ import '../src/Styles/App.css';
 import React from 'react';
 import {Route, Routes} from "react-router-dom";
 import LoginWelcome from './Login/Login-Welcome';
-import UserMainPage from './UserMainPage';
-import Twitter from './Tweets';
-import OtherUserPage from './OtherUserPage';
-import UserMainPageFav from './UserMainPage-Fav';
-
-
-
-
+import UserMainPage from './UsersPages/UserMainPage';
+import Twitter from './Tweets/Tweets';
+import OtherUserPage from './UsersPages/OtherUserPage';
+import UserMainPageFav from './UsersPages/UserMainPage-Fav';
 
 function App() {
   
  return (
-    <div className="App"> 
-    
-    <Routes>          
-         <Route exact path="/" element= {<LoginWelcome/>}/>               
-        <Route path="/twitter" element= {<Twitter/>}/>
-        <Route path="/UserMainPage" element= {<UserMainPage/>}/>
-        <Route path="/UserMainPageFav" element= {<UserMainPageFav/>}/>
-        <Route path="/OtherUserPage" element= {<OtherUserPage/>}/>
+    <div className="App">     
+      <Routes>          
+          <Route exact path="/" element= {<LoginWelcome/>}/>               
+          <Route path="/twitter" element= {<Twitter/>}/>
+          <Route path="/UserMainPage" element= {<UserMainPage/>}/>
+          <Route path="/UserMainPageFav" element= {<UserMainPageFav/>}/>
+          <Route path="/OtherUserPage" element= {<OtherUserPage/>}/>
       </Routes>
-     </div>
+    </div>
   );
 }
